backend/requirements.txt
asgiref==3.5.2
Django==3.2.6
djangorestframework==3.12.4
django-cors-headers==3.13.0
django-debug-toolbar==3.2.4
django-filter==2.4.0
djoser==2.2.0
drf-extra-fields==3.1.1
# gunicorn==20.1.0
pillow==9.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.3
python-dotenv==1.0.0
pytz==2022.2.1
reportlab==3.6.11
requests==2.28.1
sqlparse==0.4.3
